requirements.txt
mcp[cli]>=1.0.0
requests>=2.31.0
edgegrid-python>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
